changed to using environment variables for host and port

+++ app.py
@@ -40,6 +40,9 @@ ckeditor = CKEditor(app)
 load_dotenv()
 database_url = os.environ.get('DATABASE_URL')
 secret_key = os.environ.get('SECRET_KEY')
+host = os.environ.get('HOST')
+port = os.environ.get('PORT')
+debug = os.environ.get('DEBUG')
 
 app.config['SQLALCHEMY_DATABASE_URI'] = database_url
 
@@ -156,30 +159,30 @@ def add_user():
 @app.route('/user/delete/<int:id>')
 @login_required
 def delete_user(id):
-	# Check logged in id vs. id to delete
-	# if id == current_user.id:
-	user_to_delete = Users.query.get_or_404(id)
-	name = None
-	form = UserForm()
+    # Check logged in id vs. id to delete
+    if id == current_user.id:
+        user_to_delete = Users.query.get_or_404(id)
+        name = None
+        form = UserForm()
 
-	try:
-		db.session.delete(user_to_delete)
-		db.session.commit()
-		flash("User Deleted Successfully!!")
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User Deleted Successfully!!")
 
-		our_users = Users.query.order_by(Users.date_added)
-		return render_template("add_user.html", 
-		form=form,
-		name=name,
-		our_users=our_users)
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template("add_user.html", 
+            form=form,
+            name=name,
+            our_users=our_users)
 
-	except:
-		flash("Whoops! There was a problem deleting user, try again...")
-		return render_template("add_user.html", 
-		form=form, name=name,our_users=our_users)
-	# else:
-		# flash("Sorry, you can't delete that user! ")
-		# return redirect(url_for('dashboard'))
+        except:
+            flash("Whoops! There was a problem deleting user, try again...")
+            return render_template("add_user.html", 
+            form=form, name=name,our_users=our_users)
+    else:
+        flash("Sorry, you can't delete that user! ")
+        return redirect(url_for('dashboard'))
 
 
 
@@ -794,7 +797,8 @@ if __name__ == '__main__':
     # db.create_all()
     # app.run(debug=True)
 
-    app.run(host='0.0.0.0', port=3000)
+    # app.run(host='0.0.0.0', port=3000)
+    app.run(host=host, port=port, debug=debug)
     # app.run(debug=True, port=os.getenv("PORT", default=3000))
